requirements.txt
stl
click

# Maybe
path.py

pytest
